# specs/001-rpg-social-media/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: RPG Social Media API
  description: |
    Twitter-like social media platform with RPG gamification mechanics.
    Users earn XP through interactions, level up, and unlock features progressively.
  version: 1.0.0
  contact:
    name: RPG Social Media Team
servers:
  - url: https://api.rpg-social.workers.dev
    description: Production (Cloudflare Workers)
  - url: http://localhost:8787
    description: Local development (wrangler dev)
security:
  - cookieAuth: []
...
paths:
...
  /api/posts/{postId}/like:
    post:
      tags: [Interactions]
      summary: Like a post
      description: |
        Awards XP:
        - Liker: 1 XP
        - Post creator: 2 XP
      security:
        - cookieAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  like_count:
                    type: integer
                  xp_awarded:
                    type: object
                    properties:
                      liker:
                        type: integer
                        example: 1
                      creator:
                        type: integer
                        example: 2
                  level_up:
                    type: boolean
        '400':
          description: Already liked this post
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Rate limit exceeded (50 likes/hour)

    delete:
      tags: [Interactions]
      summary: Unlike a post
      description: Removes like but does not deduct XP
      security:
        - cookieAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Like removed
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/posts/{postId}/comments:
    post:
      tags: [Interactions]
      summary: Comment on a post
      description: |
        Awards XP:
        - Commenter: 5 XP
        - Post creator: 3 XP
      security:
        - cookieAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 500
                  example: "Great post!"
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment:
                    $ref: '#/components/schemas/CommentWithAuthor'
                  xp_awarded:
                    type: object
                    properties:
                      commenter:
                        type: integer
                        example: 5
                      creator:
                        type: integer
                        example: 3
                  level_up:
                    type: boolean
        '400':
          description: Content exceeds 500 characters
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Rate limit exceeded (20 comments/hour)
  /api/users/{username}/follow:
    post:
      tags: [Interactions]
      summary: Follow a user
      description: Awards 5 XP to the followed user
      security:
        - cookieAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User followed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  followers_count:
                    type: integer
                  xp_awarded_to_followee:
                    type: integer
                    example: 5
        '400':
          description: Already following this user or trying to follow self
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Interactions]
      summary: Unfollow a user
      description: Removes follow but does not deduct XP
      security:
        - cookieAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Unfollowed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Feed ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
...
  schemas:
...
    CommentWithAuthor:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        post_id:
          type: string
        content:
          type: string
        created_at:
          type: integer
        author:
          type: object
          properties:
            username:
              type: string
            level:
              type: integer
            avatar_url:
              type: string
              nullable: true
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...